# article2tweet/openapiMedium.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Medium API (Unofficial)",
    "version": "1.0",
    "x-twitter": "https://twitter.com/medium_api",
    "x-linkedin": "https://www.linkedin.com/company/medium-api",
    "x-logo": {
      "url": "https://mediumapi.com/imgs/logo/logo-dark-color.svg"
    },
    "description": "Medium API (Unofficial) helps you to quickly fetch data from Medium's Website (https://medium.com). \n\nYou can gather data related to users, publications, articles (including its textual content), latest posts & top writers within a topic/niche, etc…\n",
    "contact": {
      "name": "Nishu Jain",
      "email": "nishu@mediumapi.com",
      "url": "https://mediumapi.com"
    },
    "termsOfService": "https://mediumapi.com/terms_of_use.html"
  },
...
  "servers": [
    {
      "url": "https://medium2.p.rapidapi.com"
    }
  ],
  "paths": {
...
    "/user/{user_id}": {
      "get": {
        "summary": "Get User Info",
        "tags": [
          "User"
        ],
        "description": "Returns user-related information such as their\n\n- Username\n- Full name\n- Bio\n- Tier (FREE, MEMBER or FRIEND)\n- Followers count\n- Following count\n- Publication Following count\n- Twitter Username\n- Profile Image URL\n- Background Image URL\n- Logo Image URL\n- Tipping Link\n- List of tags in which the user is Top Writer\n- When did the user become a Medium Member (conditional)\n- When did the user become \"Friend of Medium\" (conditional)\n- Whether the user has Lists (boolean)\n- Whether the user is a Medium-verified book author (boolean)\n- Whether the user is enrolled in Medium Partner Program (boolean)\n- Whether the user has allowed notes on their articles (boolean)\n- Whether the user is suspended or not (boolean)\n\n**Note:** If you don't know the `user_id`, you can get it from the _Get User ID_ endpoint (`/user/id_for/{username}`).\n",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "description": "It's a unique hash id assigned to every single Medium user.",
            "schema": {
              "type": "string",
              "example": "1985b61817c3"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "User ID",
                      "example": "1985b61817c3"
                    },
                    "username": {
                      "type": "string",
                      "description": "An unique name chosen by every single Medium user.",
                      "example": "nishu-jain"
                    },
                    "fullname": {
                      "type": "string",
                      "description": "Full name of the user.",
                      "example": "Nishu Jain"
                    },
                    "bio": {
                      "type": "string",
                      "description": "A short description about the user.",
                      "example": "Obsessed with Tech Biz Arts & Words; Does NOT dumb down the writing; Skilled Wordsmith; Delivers the best"
                    },
                    "tier": {
                      "type": "string",
                      "description": "Tier of the user. It can be FREE, MEMBER or FRIEND.",
                      "example": "FRIEND"
                    },
                    "top_writer_in": {
                      "type": "array",
                      "description": "List of tags in which the user is a Top Writer.",
                      "items": {
                        "type": "string",
                        "example": "programming"
                      }
                    },
                    "followers_count": {
                      "type": "integer",
                      "format": "int32",
                      "description": "Number of followers the user has.",
                      "example": 450
                    },
                    "following_count": {
                      "type": "integer",
                      "format": "int32",
                      "description": "Number of authors the user is following.",
                      "example": 4
                    },
                    "publication_following_count": {
                      "type": "integer",
                      "format": "int32",
                      "description": "Number of publications the user is following.",
                      "example": 2
                    },
                    "twitter_username": {
                      "type": "string",
                      "description": "Twitter (X) username of the user.",
                      "example": "one_anachronism"
                    },
                    "is_writer_program_enrolled": {
                      "type": "boolean",
                      "description": "Whether the user is enrolled in the Medium Partner Program.",
                      "example": true
                    },
                    "is_suspended": {
                      "type": "boolean",
                      "description": "Whether the user is suspended or not.",
                      "example": false
                    },
                    "has_list": {
                      "type": "boolean",
                      "description": "Whether the user has created Lists.",
                      "example": true
                    },
                    "is_book_author": {
                      "type": "boolean",
                      "description": "Whether the user is a Medium-verified book author.",
                      "example": false
                    },
                    "allow_notes": {
                      "type": "boolean",
                      "description": "Whether the user has allowed notes on their articles.",
                      "example": true
                    },
                    "medium_member_at": {
                      "type": "string",
                      "description": "Date when the user became a Medium Member.",
                      "example": "2020-06-24 16:05:46"
                    },
                    "friend_since": {
                      "type": "string",
                      "description": "Date when the user became \"Friend of Medium\".",
                      "example": "2024-01-10 12:30:21"
                    },
                    "image_url": {
                      "type": "string",
                      "description": "Profile image of the user (URL).",
                      "example": "https://miro.medium.com/1*C92Hx7k9nRM7TPlrmhgW9w.jpeg"
                    },
                    "logo_image_url": {
                      "type": "string",
                      "description": "Logo image of the user (URL).",
                      "example": "https://miro.medium.com/1*jp6zL2GND2FnDy1WGZOdpQ.png"
                    },
                    "bg_image_url": {
                      "type": "string",
                      "description": "Background image of the user (URL).",
                      "example": "https://miro.medium.com/1*5EKkWUnrRoveFz88dcWP4Q.png"
                    },
                    "tipping_link": {
                      "type": "string",
                      "description": "A link where you can tip the Medium author.",
                      "example": "https://www.paypal.com/paypalme/sanskarjain1997"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/{user_id}/articles": {
      "get": {
        "summary": "Get User's Articles",
        "tags": [
          "User"
        ],
        "description": "Returns a list of `article_ids` written by a specified user,  with a maximum of 250 `article_ids` per request. \n\nIf the user has published over 250 articles, use the `next` query string  parameter in subsequent API calls to retrieve additional articles.\n",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "description": "Unique hash id assigned to every single Medium user.",
            "schema": {
              "type": "string",
              "example": "6e2475a6e38a"
            }
          },
          {
            "name": "next",
            "in": "query",
            "required": false,
            "description": "Use this parameter to get more results on the next page.",
            "schema": {
              "type": "string",
              "example": "1625519209064"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "User ID",
                      "example": "6e2475a6e38a"
                    },
                    "associated_articles": {
                      "type": "array",
                      "description": "List of article IDs associated with the user.",
                      "items": {
                        "type": "string",
                        "example": [
                          "81adea1a7973",
                          "f62685b42923",
                          "5105fc08e749",
                          "6c2d076d2e82",
                          "6fe20446d139"
                        ]
                      }
                    },
                    "count": {
                      "type": "integer",
                      "description": "Number of articles returned.",
                      "example": 5
                    },
                    "total_pinned_articles": {
                      "type": "integer",
                      "description": "Total number of articles pinned by the user.",
                      "example": 2
                    },
                    "next": {
                      "type": "string",
                      "description": "Use this parameter to get more results on the next page.",
                      "example": "1625519709163"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-rapidapi-key"
      }
    }
  }
}